Updated the Sound class so it has a new set method which also takes as an argument the number of CPU cycles elapsed since the last TIA sound update. The default implementation of this method invokes the old set method for backwards compatibility with derived classes.
This new set method will be used in the 1.3 release to improve the quality
of the sound emulation (e.g., to support games like Pitfall II which update
the sound registers many times during a frame).

The TIA class was updated to invoke the new set method of the Sound class.
Finally, the TIA pauseState data member was renamed to myPauseState to be
consistent with the other data members.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@55 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba