Fully updated X11 and SDL versions to support state loading and saving. Appropriate error messages are printed to the screen.
Changed location for rc and state files for X11 and SDL versions.
When Stella starts, it will create the directories '.stella' and
'.stella/state' under $HOME, if they don't already exist.

All state files will be placed in '$HOME/.stella/state'.  The locations
for the rc and properties file have also changed.  They are now
found in '$HOME/.stella/stellarc' and '$HOME/.stella/stella.pro'.

Changed shading of the screen when pause is pressed.  Now the colors
are set to 75% of normal, not 50% (as before).  I think this looks
better.  Also, shading is no longer performed in 8-bit modes, since
there weren't enough colors to do it anyway.

Made joystick code optional again in the X11 and SDL versions.  This
is in preparation for a code merge (in the SDL version) for the
Zaurus PDA.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@101 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba