Created a new Sound class that does almost nothing :)  Useful for compiling (or running) Stella with no sound support whatsoever.
Added ability to compile multiple sound backends into the X11 and SDL
versions, and to dynamically select from among them at run-time.  For
now, the X11 version can't use the SDL sound core.  This is in
preparation for the ALSA sound backend, which may or may not make it
into Stella 1.3.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@144 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba