Initial support for key remapping is finally here.  Now for the bad news:
1) It exists in the core, but only the SDL port is updated to use
it.  And the other ports don't even compile yet ...

2) The mappings for keyboard and joystick are saved in the rcfile, but
for reasons of efficiency and ease of programming, the format is somewhat
cryptic.  But the plan is to never have to edit the maps directly anyway.

3) There is no GUI in Stella to set the remapping.  This is the next
major thing to do, after I get the other ports to at least compile.

4) I haven't documented the format of the maps in the rcfile.  Its not
difficult to modify, just tedious.  If anyone is reading this and really
wants to know (or even cares), email me directly for instructions.

Enjoy :)


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@178 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba