Extended joystick enumerations.  The core event handling now supports 8 joysticks with 20 buttons (and 2 axis) each.  That should be more than enough for now (famous last words).
Of course, this doesn't mean that whatever system you run Stella under
can have 8 joysticks with 20 buttons.  It means that *IF* your system
does have a 20-button joystick, then Stella can handle it without
a recompile.

This was mainly included for those people that have gamepads with
hats, throttles, etc.  There shouldn't be more than 20 of those
per gamepad, so each would be mapped to a button.

For now, the event handling core only knows about buttons and
left/right, up/down axis events (which can be considered a button
as well).  It will be up to the frontends to map whatever your
gamepad has (hat, wheel, etc.) to something that Stella can
understand (a button).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@196 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba