Cleaned up the UserInterface class a little.
Added arrows for user feedback in the key remapping section of the
UserInterface class.  Still have to indicate which key is being
remapped, though.

Added another menu to the main menu (of the GUI) called
'Character Set', showing the font being used.  I found it useful
for determining what particular character to use when coding
the GUI, but I don't know if it will remain in the final release.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@197 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba