Added mouse support to the DirectInput class, and detection of mouse events to MainWin32.
Removed the DI_Event type structures.  While it made the code
look nicer, we really didn't need the performance hit (however
small it may have been) of serializing and deserializing the
events.

Removed the rest of the framebuffer code from MainWin32.
Made sure the exit event (Escape) was sent to the core,
so that current settings could be saved.

Next I'll look at the joystick code (should be quite easy).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@208 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba