Some performance tweaks for the SDL OpenGL port.  For reference, I'm on a Celeron 766 with a crappy ATI card.  Under normal usage, CPU usage has dropped from 60% to 8.5%.  But there is still a slowdown when showing alpha-blended rectangles at the current framerate.  I'm about to mark this up as being caused by the crappy card and OpenGL implementation I'm using, so performance may not get much better than this.  On my development system (even before this latest patch), CPU usage never goes above 14% (granted, that system is a P4-2.4 with GeForce4MX card).
I'll try to run on a similar system to this Celeron, but with
an NVidia card, and see if the ATI card drivers are actually
at fault.

But I really do think its the OpenGL drivers, since software
mode never uses more that 4% of the CPU, and that is at 1280x1024.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@212 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba