Broke the dependency between the FrameBuffer and Sound classes wrt to the MediaSource.
Previously, in the main run loop for a port, you had to call
FrameBuffer::update and then
Sound::updateSound(FrameBuffer::MediaSource).  Now, you can call
FrameBuffer::update() and then Sound::update().  But since the
ordering here is important, I've introduced a Console::update()
that does it all.  So the main run loop for each port just
got a bit simpler.

Also changed around the Snapshot class a bit.  All this is to
reinforce the fact that the MediaSource is owned by the core,
and the FrameBuffer, Sound, and Snapshot classes make use of
it.  Previously, it seemed as if the FrameBuffer owned the
MediaSource.

I'm finding all these little inconsistencies because of
writing the Porting.txt document, and trying to explain
why something is being done a certain way (and if I
can't explain it, I rewrite it).  So if for no other
reason, the Porting.txt document has already been
valuable.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@214 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba