StellaX now loads metadata from the stella.pro file (Filename, Manufacturer, Rarity, Note, etc).
Since the loading of the metadata is slow, I added a cachefile
for the gamelist (named 'stellax.cache').  If this file is present,
the gamelist is created based on the contents of the cachefile.
This is quite fast, since no stella.pro parsing or ROM loading
is done.  The cache file is recreated each time a full reload is
done.

Pressing the 'Reload' button always forces a full reload (not from
cache), so this is useful if you change the ROM directory (or add
ROMs).  Still TODO is add a check to see if the ROM dir has had
any files added to it since last time you ran StellaX, and do a
full reload.

Got rid of the filename field in the game listview.  Now only show
'Cartridge Name', 'Cartridge Manufacturer' and 'Cartridge Rarity'.
If some game doesn't have an entry in the stella.pro file, its
filename will be shown as cartridge name.

Still TODO are add sorting/permanent resizing options to the column
headers in the listview, and add options to the Config Page.  It's
getting close, folks.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@300 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba