Added first pass of the ROM launcher.  When you press 'Start' in the ROM launcher, it launches Frostbite.  Then pressing 'Escape' goes back to the launcher, and you're able to then launch Frostbite again!  Success !!! Still TODO is actually get a ROM listing and use the selected game, but the fact that it works multiple times for some game means that the infrastructure is working correctly :)
Changed behaviour of the 'Escape' key.  Specifically, it now only acts
as a key to enter ROM launcher mode.  In the case where the emulation
was started without the launcher, the key will now do nothing.  From now
on, the only way to quit Stella is Ctrl-Q (or equivalent for OSX),
close the window, or from the 'Quit' button.

Also, the 'Enter launcher mode' event will eventually be made remappable,
as will the 'Enter menu mode' event (currently the 'Tab' key).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@412 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba