Cleaned up the LauncherDialog a little.
Removed the saving of TIAbit on/off into the state files.  These
are debugging tools, and aren't really part of a state file (and hence
shouldn't be saved with a state file).  Now, whenever a state is
(re)loaded, the TIABits are all turned back on.  That means the state
files are still compatible with Stella 1.4.2.

Added "Enter/exit menu mode" and "Enter launcher mode" to the remappable
events.  They default to 'Tab' and 'Escape', respectively.  While you
can add new mappings for these events, those defaults can't be removed
(otherwise, it would be possible for a user to disable menu mode, and
never be able to get back into the menu to fix the problem).

Fixed some problems with the overall state machine.  Sometimes the cursor
was hidden when it shouldn't have been, and vice versa.

Fixed crash when toggling between software and OpenGL while in ROM
launcher mode.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@413 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba