Made LauncherDialog look for a system-specific gamelist cache file.
Modified the framerate argument, in that it can only be changed from the
integrated GUI, and only for the current game.  The framerate is now
properly set when starting a new game, or when switching between NTSC/PAL
mode, and is based on the ROM format.  The ability to change the framerate
may disappear entirely before the next release (I see no real need for it
anymore).

Some minor code cleanups.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@419 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba