Scratch another most-requested feature of the list.
Just added support for zipped ROMs.  Right now, it assumes that the
first file in a zipped archive is the ROM (if it's a multi-file archive).
I thinks that's the way I'll leave it, since otherwise I need another way
to determine which one is the ROM, and that means looking at extensions
or something (which is something I *don't* want to do).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@428 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba