Fixed inttypes support so that Stella compiles in Visual C+ .Net (which is currently the lowest supported version of Visual Studio that Stella can be compiled in).  It seems that VS.Net already knows about the inttypes, so an explicit 'include <inttypes.h>' isn't required.  However, when compiling under MingW, the include *is* required.
MingW is still the preferred environment for compiling Stella though, as
VS.Net doesn't read the SDL environment variables.  So if you're working
on Stella, feel free to use VS.Net (as long as you don't use anything
specific to that environment).  But all binary releases of Stella will use
MingW, and code that doesn't work in MingW will be reworked/removed.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@435 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba