Removed 'mergeprops' commandline argument.
Added Alt-s key to merge current properties into stella.pro.  Together with
the Ctrl-s key which saves current properties to a new properties file,
there was no longer any need for the mergeprops argument.

Major update of the Stella manual, including the further consolidation
of the main codebase with the OSX version, as well as snapshots of the
integrated GUI.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@453 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba