Made the Debugger dialog use a TabWidget again, instead of it being another Dialog.  This was necessary since it seems Dialogs can't contain other dialogs in the GUI core.
Added keyboard navigation to the TabWidget class.  For now, it's only
enabled for the Debugger.  Ctrl-Tab/Shift-Ctrl-Tab switches to the
next/previous tab (respectively), and Tab/Shift-Tab switches to the
next/previous widget in the current tab.

There's still a bit of work to do, but the infrastructure is there.
That's it, I'm taking the night off, because this deceptively simple
concept has taken about a day to implement ...


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@479 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba