Re-added (center) commandline argument, which was in a previous version of Stella.  It defaults to true/on, so if you don't want centering, use '-center false' on the commandline at least once (or edit the config file).
Some more work on the RamWidget and ByteGridWidget.  Entering values in hex
now works correctly, and 128 byte RAM area is now updated.

More work on the automatic updating of the tabs.  Tabs are now updated when
debugging mode is entered, or when a new tab is selected.

Shortened the width of the debugging tab area and added an area for commonly
use actions/buttons on the right.  Currently it contains 'Step' and 'Trace'
buttons.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@517 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba