Implemented prefix operators < and > (low byte and high byte) in CLI. They work just like the ones in DASM: "<1234" is "34" and ">1234" is "12". They can of course be used with labels, and they can be prefixed with "*" for dereference.
Implemented binary input in CLI. Like DASM: prefix with "%" to treat
your argument as binary. "%10101010" means hex "ff". Can be stacked with
other prefixes.

Stacking order of prefixes is important. "*" must come first, if present.
Then ">" or "<" if present (but not both!). Then "%" if present. Currently
there's no real error checking for this.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@528 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba