Fixed some Valgrind errors, mostly non-initialized data.
Fixed some memory leaks in EquateList, and moved to using a dynamic array.
This greatly simplifies the code and abstracts away all new/delete
operations.  More cleanup can still be done, since the symfile no longer has
to be scanned for # of lines.  Still TODO is similar code for the watchlist
stuff, so all memleaks can be eliminated.

Changed launcher so that if Stella is started with no romdir specified
(maybe first time it's been used), the launcher options are shown so
you can select the romdir.  Thanks to Brad for the advice.

Updated LauncherOptionsDialog to send a signal when a romdir has been
set.  When that happens, the launcher automatically reloads the rom listing.
Again, thanks to Brad for the advice.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@539 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba