Label completion behaves almost identical to bash completion now. If you have labels "food" and "foobar", and no others starting with "f", and you type "pr f<TAB>", it will show you both of them, then leave your prompt looking like "pr foo" with the cursor after the 2nd "o".
Differences from bash:

- In case of multiple completions, bash's default behaviour requires you
  to press Tab twice to see them. We only need one Tab. bash can actually
  be configured to only need one Tab, too.

- We don't beep. Ever.

- We're case-insensitive. bash is case-sensitive by default, though it
  can be configured to be insensitive.

There's one remaining bug in the completion: when we get to the bottom of
the scroll buffer, and the buffer is full (e.g. when the scroll bar handle
is at its minimum size), *and* there's only one possible completion, the
console prints a new blank line after the new (completed) prompt.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@559 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba