Second pass at configure support; it actually seems to work this time. I've only tested it in 32bit Linux, but I don't forsee any problems. Dependencies seem to work correctly, recompiling parts of the codebase as necessary.
New files added to the codebase must be listed in the 'module.mk' file
in the same directory where the new file is placed.  Other then that,
no configuration is required.

All configurable options are by default enabled.  They can be manually
disabled by an appropriate command to the configure script, but the
script is smart enough to test for the required libraries and disable
support when one isn't found.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@579 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba