3E and 3F bankswitching now pass pokes to addresses $00-$3F through to the TIA. For 3E, this means game authors don't need to access the TIA through a mirror (and there's no reason they should on real hardware, either). It also means Armin's test cart image works.
3F ROMs still need to use a mirror: poking to $00-$2C writes to the
TIA *and* switches banks, which is not that you'd want to do. Stella's
behaviour now matches a real Tigervision 3F cart, for what it's worth.

Unfortunately I had to do something nasty to allow the Cart classes to
directly access the TIA: see System.hxx and System.cxx (search for
the tiaPoke() method.)


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@627 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba