Totally revamped DebuggerDialog.  The dialog now takes up the whole screen, which means that the area right of the TIA image can be used for widgets (that's the next thing I'll be doing).
Created TiaOutputWidget, which is contained in the DebuggerDialog
and handles all drawing/updating of the TIA image while in
debugger mode.  Some advantages of this are a greatly simplified
FrameBuffer::update() wrt debugger mode and instant updates for frame
and scanline advance.  Another big advantage is that the TIA image
is now a widget, which means it can receive key/mouse events.  These
events can trigger updates in other parts of the interface (imagine
clicking on a certain line and having scanlines filled to that
area, etc).  The biggest disadvantage right now is that the TIA is
redrawn each time the dialog changes (ie, when a key is pressed).
This needs to be optimized.

Fixed bug whereby toggling video modes while in the debugger
sometimes screwed up the palette.

Removed some FIXME's, since the functionality has been provided.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@678 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba