Added RomListWidget class, which derives from CheckListWidget.  Created a new class since the RomWidget isn't just a checkbox and string. Eventually, it will consist of multiple fields.
Added highlighting of the line at current PC to RomWidget.  It attempts
to reduce scrolling as much as possible (since each scroll requires a
re-disassembly) by instead scrolling the line at PC.  When the bottom
of the list is reached, it 'snaps' back again to the top.  So, for
example, moving through 31 consecutive items in a list of 30 only scrolls
once, not 30 times.  This is different from normal
ListWidget/CheckListWidget operation, which scrolls the whole list
line by line at the bottom of the current view.

One can scroll around in the RomWidget, and then press 'space' to snap
back to the line of current PC (so you'll never get lost while
scrolling through code).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@748 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba