Massive rearrangement of most GUI-related classes (sorry Mark for the changes you'll have to make to the XCode project).
The '--disable-developer' option to configure now works properly,
and developer/debugger support can be completely disabled.  When
disabled, the code will not be compiled at all, which gives an approx
35% reduction in code size (useful for the PSP port).

Cheat support is currently broken as it depends on a debugger that
may not be present, but this will be fixed soon (it isn't near
finished yet anyway).

Made changing of game properties and toggling TIA bits not exclusively
a developer thing, since there's no reason why anyone can't use it.

Bumped version to beta1.  Stella has been stewing for the past 6 months,
and it's almost about time for a release.  I promised last time that
releases would be more frequent, and I've already broken that promise :)
That means some features of the debugger won't be included for next
release, but something has to be left out ...


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@751 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba