Implemented 'Fill to scanline' in TiaOutputWidget.  Currently, you have to be in 'scanline' mode for it to work (which means you must have previously selected 'Alt l' or pressed the 'Scanline+1' button). This is similar to the bug where stepping shows partial changes only when you've first selected scanline mode; this bug will be fixed shortly.
Fixed coordinate issues in TiaZoomWidget and when setting position
from TiaOutputWidget.  The position selected in TiaOutputWidget
now becomes the center point in the zoomed image.

Zooming in TiaZoomWidget now operates on the center coordinate.  That
means you actually zoom in on the center of the image (vs. before,
where you zoomed on the upper-left area).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@757 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba