Removed some redundant code in FrameBufferGL.
Made sure emulation screen height can always accommodate the menus
placed in it.

Added check when switching to PAL mode, so that the emulation height
will be a default for PAL (250), not for NTSC (210).

Added ability to switch between stereo and mono sound dynamically
from the AudioDialog ('Stereo Mode') and the commandline ('-channels').
This setting is not saved to stellarc, so if you want it to be
permanent for some game, save it to the properties file.

Added 'Cartridge.Sound' property to the game properties, and have it
default to 'Mono'.  Setting it to 'Stereo' will give stereo sound
(2 sound channels used).  Related to this, update stella.pro
'Skeleton+' entry to use stereo sound.

Correctly update the sound framerate when switching between NTSC/PAL,
since the new sound core knows what to do in such a case.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@768 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba