Some minor tweaks which will hopefully speed things up for the GP2X port.
Resurrected the low-compatibility 6502 CPU code, since it's supposed
to be somewhat faster.  This is activated internally by GP2X by
settings "cpu" to "low".  This can also be set from the commandline,
but I'm not going to document the fact.  All systems that can handle
hi-compat mode should be using it.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@999 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba