First pass at making the GUI font-sensitive.  Except for the debugger, everything should look exactly the same.  This infrastructure will eventually allow a set of fonts to be included in Stella, and for the user to select a bigger or smaller font and have the GUI automatically resize to it.
At some point, the ability to zoom the ROM launcher and debugger
will be removed, and will be replaced by selection of different font
sizes.  As well, zooming of the MediaSource will be converted to a
graphics filter (scale2x/3x/4x), and others such as hq2x or mame2x
will also be available.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1009 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba