Removed all references to stella.pro locations from configure and Makefile scripts, since the properties are now built in, and will never need a global storage directory.
Reworked the ROM launcher again, by moving the 'Reload' button to the
Launcher options dialog (by popular request from the GP2X forums).  The
'Go Up' button is now always present on the main launcher interface, but
is disabled when not in 'rom browse' mode.

Added a button to VideoDialog for the recently added 'tiadefaults'
functionality.  It defaults to off, which is the current behaviour in
Stella.  (De)selecting this item will require a ROM reload.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1058 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba