Major update to the palette support.  Internally, all palette info now belongs to the Console class.  A single user definable palette can now be automatically loaded.  This new palette file must be named 'stella.pal', and located in the app base directory ($HOME/.stella for Linux, installation dir for Win32).
The new palette file must be at least 768 bytes (only the first 768
bytes are used).  The first 384 bytes are 128 red-green-blue 8-bit
values, or 24-bits per pixel for NTSC.  The next 384 bytes/128 colours
are the same, except for PAL.  The PAL colour-loss palette is automatically
calculated from these, so it's not necessary to do so in the palette file.

Added 'user' to the list of options that can be specified for '-palette',
and also added the option to the VideoDialog.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1154 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba