Added some OpenGL options to FrameBufferGL that only exist on SDL 1.2.10 or greater.  I'm hoping that one older systems, the code will still run and just ignore these new options.
Added commandline argument '-gl_vsync', which enables v-sync'ed updates
for OpenGL, if it's available (see above).  Also added a setting for
this in the VideoDialog.

Fixed OSX OpenGL handling so that it doesn't specify an OpenGL library name,
and updated FrameBufferGL to use the default name in that case.  OSX users
will probably need to reset their settings file for this to take effect.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1175 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba