Disable some of the UI controls for the wince port. We don't want the user disoriented with options that are: 1) ignored byt the port, 2) Cannot fit in small screen devices.
As far as the #ifdef badness goes this is not too bad. The conditional compiled code blocks are scarce. In my opinion they do not justify an architectural approach, which would mean adding methods for enabling/disabling widgets.

Also note in OptionsDialog.cpp we were leaking memory by instantiating twice the InputDialog.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1301 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba