Fixed 'Game Info' button in Options Dialog to only be shown when a valid ROM has been selected in the ROM launcher (or if we're in emulation mode).
BTW, I forgot to mention in the previous commit that Stella now only
recognizes ROMs if they're named with the following extensions: .bin,
.a26, .rom, .zip, .gz (not case sensitive).  Right now, attempting to
select something else prints an error message.  Eventually, a MessageBox
will be shown indicating that the filename isn't valid.

For now, this restriction is only for files started from the built-in
launcher.  I haven't decided if it should apply to the commandline as well.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1427 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba