Reactivated 'mouse is paddle X' functionality by integrating it directly into the Paddles class.
Removed '-paddles' commandline argument.  You can still change the paddle
that the mouse is emulating from the UI or with Ctrl 0..3, but its no
longer a setting that can be saved.

Added message to LauncherDialog when a ROM is invalid and can't be loaded.

Reworked the GameInfoDialog a little wrt swapping console ports.  Behind
the scenes, the Console.SwapPorts property is still being used, but the
UI is now more consistent (you no longer select the left and right
controller, but the controllers for players 0 and 1, and which port
they'll be plugged into).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1435 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba