Changed my mind again, and decided to reactive the ROM Info stuff; it's just too cool a feature not to include in the next release ;)
Checked in some updated UI font-aware code.  All of the UI's are not
converted yet, but for the next release the only one that really matters
is LauncherDialog.

Added '-launcherfont' commandline argument and associated UI config
to set the size of the font used in the ROM launcher (for now, small
or large).  This is useful if you activate the ROM Info and can't
zoom the entire UI by 2x.  Previously, this meant the ROM listing font
was very small, but now you can have the launcher in 1x mode with a
larger font *and* the ROM info stuff also visible.  More changes are
coming in a future release, so this will have to do for now.

Reworked UIDialog to contain tabs, since the options were starting to
become unwieldy.

This pretty much ties up the last remaining items for this release,
so after a little testing and updating the manual, it's time for 2.5.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1438 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba