Made TIA auto-frame detection kick in every 32 frames instead of 64.
Further tweaking to the RIOT timer values.  I reverted to the z26
'0x40000' number, even though I'm not entirely clear why it's being
used.  Without it, several PAL ROMs fail in an infinite loop of
reading from INTIM and checking for zero.  I really hope that's the
last of the RIOT stuff, since I've been looking at it for the past
two months (the AVox and SaveKey issues started there as well).


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1531 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba