Scaled back some TIA changes from the previous alpha release.  The changes are too invasive to fix right now, and I haven't researched what the correct route should be.  Basically, Stella doesn't properly emulate the behaviour of writes to RESxx and NUSIZx while graphics are currently being drawn.  The tweaks currently present restore functionality to the 2.8.4 release, but are still incorrect in some ROMs (Bumper Bash and Pole Position).
Reworked 'object disable' mode in the TIA class.  It should now be
slightly faster, and correctly disable objects in all cases
(previously, it was possible for some objects to be enabled depending
on the state of other objects, even when they were specifically
disabled).

Changed NUSIZx descriptors in the debugger to more clearly indicate
what's going on.

Cleaned up TIATable enum's, and eliminated duplication enumerations in
different parts of the codebase.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1855 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba