Fixed scanline => framerate calulation bug for those ROMs that generate many more scanlines than a real TV would allow.  This fixes both graphical corruption and slowdowns in Q-Bert ROM (which contains code to do a tight loop extending VBLANK time, in effect generating a 'frame' with 1000's of scanlines).
First pass at a method to determine the first 'relevant' scanline of
a TIA frame.  This isn't easy, because there are many conditions
for this to occur.  The eventual goal is to eliminate Display.YStart
tweaks in the properties file, and have the TIA code itself figure out
the best value.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1858 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba