Updated Thumb ARM emulation to log its output instead of printing to the console (which doesn't exist on every platform).  Also use proper C++ stringstreams instead of C-style arrays (which are evil).  If an exception is thrown, start the debugger with a fatal error message.  Currently, this message is simply printed in the debugger info window.  Still TODO is add a dialog to show the entire input from within the UI, as well as a way to exit the ROM when a fatal error occurs.
Cleaned up the API a little, adding explicit declarations for when a method
can throw an exception.  Also merged some empty classes directly into their
header files.

Added functionality to the System class for querying whether it has 
been reset while in 'autodetect' mode.  Autodect mode is defined as
when Stella starts and temporarily runs to autoconfigure itself.
Some classes (such as Thumb ARM emulation) need to know this, and
suppress debugging output/throwing exceptions when not running in
normal emulation mode.

I hope to have a new release done by the end of this week, or the end
of the month at the latest.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@2232 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba