Some infrastructure work for CompuMate handling.  The CM handler now disables normal keyboard input when in the debugger.  This fixes issues where commands/shortcut keys typed in the debugger would be interpreted as actual CompuMate input.
For now, in-debugger key processing is completely disabled in such a
case.  Eventually, we should probably add a virtual keyboard to the
CM RIOT area, and have it send its data directly to the CM handler.

Bumped version # slightly.  I hope to get the next release ready to
go in a few weeks.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@2705 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba