Updated PNG and ZLIB libraries to latest released versions.
Since Zlib 1.2.10, it seems that zlib and png need to be paired
together.  The UNIX 'configure' script has been updated to either use
system png and zlib libraries, or both as built-in libraries.  This
shouldn't affect Windows (which uses both built-in anyway), but it
does require changes to OSX (which currently doesn't build at all).