Add a GitHub action to build on Linux and macOS
As with Travis CI, the test is only deemed successful on each platform
if the Stella build completes successfully.

For Linux, the test runs on Ubuntu with the toolchain test build
repository, using G++ 9, and the packaged version of SDL2, which is
guaranteed to be at least 2.0.10.

For macOS, SDL2 2.0.10 is still built from the upstream source code;
however that no longer builds with the newer Xcode versions available
in GHAs, so the build now uses "Unix-style" ./configure && make.

The runners provide two cores on Linux and three cores on macOS, so
the compile stages use the appropriate -j setting to build in
parallel.

Signed-off-by: Stephen Kitt <steve@sk2.org>