New PatchRomCheat type (#1088)
* Update index.html for ROM Patching Cheats

Added description of ROM patching cheats, a new 9 digit cheat type.

* New PatchRomCheat type

Allows a type of ROM cheat with no concept of banks, it merely patches the offset of the ROM - will make it possible to add cheats I'm in the process of creating. 
Basically:
ON:
if (peek(ROM_Address)==original_data)
{
  poke(ROM_Address)=new_data
  set_cheat_to_enabled = TRUE; 
}
OFF:
if ( set_cheat_to_enabled = TRUE )
{
  poke(ROM_Address)=new_data
}
 set_cheat_to_enabled = FALSE; 

   9-digit (patch) code: aaaaddeef

   aaaa = cartridge address
   dd = new data to poke
   ee = original data (must be correct or the cheat wont enable)
   f = throwaway (set to F)

* Changed year to 2025

Fix of year (I had to use the source for 7.0 as building git main branch is broken as SDL changes have been made)